Fix usize issues and update readme (#27)
* Fix usize issues

* Readme updates

* Remove target

+++ src/msg.rs
@@ -1,12 +1,10 @@
-#![allow(clippy::field_reassign_with_default)] // This is triggered in `#[derive(JsonSchema)]`
-
 use crate::state::{Authentication, SecretContract};
 use crate::viewing_key::ViewingKey;
 use cosmwasm_std::{Binary, HumanAddr, Uint128};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[derive(Serialize, Deserialize, JsonSchema)]
 pub struct InitMsg {
     pub buttcoin: SecretContract,
     pub butt_lode: SecretContract,
@@ -26,10 +24,10 @@ pub enum HandleMsg {
         padding: Option<String>,
     },
     Show {
-        id: usize,
+        id: u64,
     },
     UpdateAuthentication {
-        id: usize,
+        id: u64,
         label: String,
         username: String,
         password: String,
